Restructuring for the database, clean architecture, and new features. CSS refactoring.
