+++ static/layouts/cards.css
@@ -7,7 +7,6 @@ button:focus, input:focus, textarea:focus {
     outline: 0;
 }
 
-/* 
 nav {
     display: flex;
     flex-wrap: wrap;
@@ -15,23 +14,7 @@ nav {
     gap: 20px;
     padding: 20px;
     min-height: 100vh;
-    max-width: 800px;
-    margin: 0 auto;
-}
-
-.blog_wrap {
-    width: 380px;
-    margin: 0;
-} */
-
-nav {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 20px;
-    padding: 20px;
-    min-height: 100vh;
-    max-width: 1250px;
+    max-width: 1050px;
     margin: 0 auto;
 }
