+++ static/layouts/content.css
@@ -1,12 +1,12 @@
 html, body {
     font-family: sans-serif;
+    box-sizing: border-box;
+    overflow-x: hidden;
+    min-height: 100vh;
     margin: 0;
     padding: 0;
     font-size: 21px;
     line-height: 1.58;
-    box-sizing: border-box;
-    overflow-x: hidden;
-    min-height: 100vh;
     background: var(--color-backg);
     color: var(--color);
 }
