+++ static/scripts/lightMode.js
@@ -9,7 +9,7 @@ lightmode.addEventListener('click', function(){
         
         document.documentElement.style.setProperty('--color-backg', '#191c20');
         document.documentElement.style.setProperty('--color', '#f7f7f7');
-        imgLightMode.setAttribute("src", "/static/image/brightness_light.png")
+        imgLightMode.setAttribute("src", "/static/assets/images/brightness_light.png")
         
         document.documentElement.style.setProperty('--card-backg', '#191919');
         document.documentElement.style.setProperty('--card-info-color', '#ffffff');
@@ -22,7 +22,7 @@ lightmode.addEventListener('click', function(){
 
         document.documentElement.style.setProperty('--color-backg', '#f7f7f7');
         document.documentElement.style.setProperty('--color', '#0e0e0e');
-        imgLightMode.setAttribute("src", "/static/image/brightness_dark.png")
+        imgLightMode.setAttribute("src", "/static/assets/images/brightness_dark.png")
 
         document.documentElement.style.setProperty('--card-backg', '#ffffff');
         document.documentElement.style.setProperty('--card-info-color', '#0e0e0e');
