+++ templates/index.html
@@ -4,10 +4,10 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0"> 
     <title>Crunchyroll RSS</title>
-    <link rel="stylesheet" href="/static/css/nav.css">
-    <link rel="stylesheet" href="/static/css/cards.css">
-    <link rel="stylesheet" href="/static/css/footer.css">
-    <link rel="shortcut icon" href="/static/image/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="/static/layouts/nav.css">
+    <link rel="stylesheet" href="/static/layouts/cards.css">
+    <link rel="stylesheet" href="/static/layouts/footer.css">
+    <link rel="shortcut icon" href="/static/assets/images/favicon.ico" type="image/x-icon">
 </head>
 
 <body>
@@ -15,7 +15,7 @@
         <div class="header-nav">
             <div class="nav-logo">
                 <a href="/">
-                    <img src="/static/image/crunchyLogoRSS.png" alt=""> 
+                    <img src="/static/assets/images/crunchyLogoRSS.png" alt=""> 
                 </a>
             </div>
             <div class="nav-button">
@@ -26,7 +26,7 @@
                 </div>
                 <div class="lightmode">
                     <button id="lightmode">
-                        <img src="/static/image/brightness_dark.png" id="imgMode" alt="lightmode">
+                        <img src="/static/assets/images/brightness_dark.png" id="imgMode" alt="lightmode">
                     </button>
                 </div>
             </div>
@@ -54,11 +54,11 @@
             <p class="text_color">created by olavodotpy</p>
         </div>
         <div class="src">
-            <a href="https://github.com/olavodotpy/AnimeNews/tree/Master">Source code/credits</a>
+            <a href="https://github.com/olavodotpy/Crunchyroll-RSS">Source code/credits</a>
         </div>
     </footer>
 
 </body>
 
-    <script src="/static/js/lightMode.js"></script>
+    <script src="/static/scripts/lightMode.js"></script>
 </html>
