+++ templates/post.html
@@ -4,10 +4,10 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{{ post.title }}</title>
-    <link rel="stylesheet" href="/static/css/nav.css">
-    <link rel="stylesheet" href="/static/css/content.css">
-    <link rel="stylesheet" href="/static/css/footer.css">
-    <link rel="shortcut icon" href="/static/image/favicon.ico" type="image/x-icon">
+    <link rel="stylesheet" href="/static/layouts/nav.css">
+    <link rel="stylesheet" href="/static/layouts/content.css">
+    <link rel="stylesheet" href="/static/layouts/footer.css">
+    <link rel="shortcut icon" href="/static/assets/images/favicon.ico" type="image/x-icon">
 </head>
 
 <body>
@@ -15,13 +15,13 @@
         <div class="header-nav">
             <div class="nav-logo">
                 <a href="/">
-                    <img src="/static/image/crunchyLogoRSS.png" alt=""> 
+                    <img src="/static/assets/images/crunchyLogoRSS.png" alt=""> 
                 </a>
             </div>
             <div class="nav-button">
                 <div class="lightmode">
                     <button id="lightmode">
-                        <img src="/static/image/brightness_dark.png" id="imgMode" alt="lightmode">
+                        <img src="/static/assets/images/brightness_dark.png" id="imgMode" alt="lightmode">
                     </button>
                 </div>
             </div>
@@ -49,10 +49,10 @@
             <p class="text_color">created by olavodotpy</p>
         </div>
         <div class="src">
-            <a href="https://github.com/olavodotpy/AnimeNews/tree/Master">Source code/credits</a>
+            <a href="https://github.com/olavodotpy/Crunchyroll-RSS">Source code/credits</a>
         </div>
     </footer>
 
-    <script src="/static/js/lightMode.js"></script>
+    <script src="/static/scripts/lightMode.js"></script>
 </body>
 </html>
